+++ src/main/kotlin/haxidenti/furrworld/player/PlayerInventory.kt
@@ -1,17 +1,17 @@
 package haxidenti.furrworld.player
 
 class PlayerInventory(
-    val items: MutableList<Int> = mutableListOf(),
+    val items: MutableList<Short> = mutableListOf(),
 ) {
     val size get() = items.size
 
-    fun add(item: Int) {
+    fun add(item: Short) {
         this.items.add(item)
     }
 
-    fun remove(item: Int) {
+    fun remove(item: Short) {
         this.items.remove(item)
     }
 
-    operator fun contains(item: Int) = item in items
+    operator fun contains(item: Short) = item in items
 }
