+++ src/main/kotlin/haxidenti/furrworld/world/Cell.kt
@@ -1,6 +1,6 @@
 package haxidenti.furrworld.world
 
-data class Cell(val floorId: Int, val objectId: Int) {
+data class Cell(val floorId: Short, val objectId: Short) {
     companion object {
         val VOID = Cell(0, 0)
     }
